backend/requirements.txt
fastapi[standard]==0.128.0
sqlmodel==0.0.31
uvicorn[standard]==0.40.0
psycopg2-binary==2.9.11
alembic==1.18.0
python-dotenv==1.2.1
passlib[bcrypt]==1.7.4
circle-developer-controlled-wallets==9.1.0
cryptography==46.0.2
requests==2.32.5
pwdlib[argon2]==0.3.0
pyjwt==2.10.1
stripe